third_party/WebKit/LayoutTests/hittesting/subframe_active_crash.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#frame:active {
  border-width: 5px;
}
</style>
<iframe id="frame"></iframe>
<script>
test(() => {
  assert_not_equals(window.eventSender, undefined, 'This test requires eventSender');

  const frame = document.getElementById('frame');
  const mouseX = frame.offsetLeft + frame.offsetWidth / 2;
  const mouseY = frame.offsetTop + frame.offsetHeight / 2;
  eventSender.mouseMoveTo(mouseX, mouseY);
  eventSender.mouseDown();
  eventSender.mouseUp();
}, 'Clicking in a subframe with :active style should not crash.');
</script>
